Renamed ocl tests' precomp files according to convention.
Also, added the missing include guard to perf_precomp.hpp.
This should fix the build.
