+++ modules/ocl/perf/perf_arithm.cpp
@@ -44,7 +44,7 @@
 //
 //M*/
 
-#include "precomp.hpp"
+#include "perf_precomp.hpp"
 ///////////// Lut ////////////////////////
 PERFTEST(lut)
 {
