+++ modules/ocl/perf/perf_blend.cpp
@@ -44,7 +44,7 @@
 //
 //M*/
 
-#include "precomp.hpp"
+#include "perf_precomp.hpp"
 ///////////// blend ////////////////////////
 template <typename T>
 void blendLinearGold(const cv::Mat &img1, const cv::Mat &img2, const cv::Mat &weights1, const cv::Mat &weights2, cv::Mat &result_gold)
