+++ modules/ocl/perf/perf_calib3d.cpp
@@ -44,7 +44,7 @@
 //
 //M*/
 
-#include "precomp.hpp"
+#include "perf_precomp.hpp"
 ///////////// StereoMatchBM ////////////////////////
 PERFTEST(StereoMatchBM)
 {
