+++ modules/ocl/perf/perf_gftt.cpp
@@ -44,7 +44,7 @@
 //M*/
 
 
-#include "precomp.hpp"
+#include "perf_precomp.hpp"
 
 ///////////// GoodFeaturesToTrack ////////////////////////
 PERFTEST(GoodFeaturesToTrack)
