+++ modules/ocl/perf/perf_matrix_operation.cpp
@@ -43,7 +43,7 @@
 // the use of this software, even if advised of the possibility of such damage.
 //
 //M*/
-#include "precomp.hpp"
+#include "perf_precomp.hpp"
 
 ///////////// ConvertTo////////////////////////
 PERFTEST(ConvertTo)
