+++ modules/ocl/test/test_arithm.cpp
@@ -54,7 +54,7 @@
 //#define PRINT_TIME
 
 
-#include "precomp.hpp"
+#include "test_precomp.hpp"
 #include <iomanip>
 
 #ifdef HAVE_OPENCL
