+++ modules/ocl/test/test_brute_force_matcher.cpp
@@ -43,7 +43,7 @@
 //
 //M*/
 
-#include "precomp.hpp"
+#include "test_precomp.hpp"
 #ifdef HAVE_OPENCL
 namespace
 {
