+++ modules/ocl/test/test_canny.cpp
@@ -43,7 +43,7 @@
 //
 //M*/
 
-#include "precomp.hpp"
+#include "test_precomp.hpp"
 #ifdef HAVE_OPENCL
 
 ////////////////////////////////////////////////////////
