+++ modules/ocl/test/test_filters.cpp
@@ -48,7 +48,7 @@
 //
 //M*/
 
-#include "precomp.hpp"
+#include "test_precomp.hpp"
 
 #ifdef HAVE_OPENCL
