+++ modules/ocl/test/test_imgproc.cpp
@@ -51,7 +51,7 @@
 //
 //M*/
 
-#include "precomp.hpp"
+#include "test_precomp.hpp"
 
 #ifdef HAVE_OPENCL
