+++ modules/ocl/test/test_match_template.cpp
@@ -43,7 +43,7 @@
 //M*/
 
 
-#include "precomp.hpp"
+#include "test_precomp.hpp"
 
 #ifdef HAVE_OPENCL
 ////////////////////////////////////////////////////////////////////////////////
